SRC/python/requirements.txt
mediapipe==0.9.3.0
opencv-python==4.7.0.68
pyserial==3.5